mod-watcher/requirements.txt
aws-cdk.core
aws-cdk.aws_lambda
aws-cdk.aws_apigateway
aws-cdk.aws_dynamodb
aws-cdk.aws_secretsmanager
aws-cdk.aws_ssm
aws-cdk.aws_iam
aws-cdk.aws_s3
aws-cdk.aws_events
aws-cdk.aws_events_targets
boto3